+++ src/screens/Home.js
@@ -1,18 +1,14 @@
 import React, { Component } from "react"
-import { View, Text } from "react-native"
-import styled from "styled-components/native"
-import Routing, { Router } from "../utilities/routing"
-
 import AppScrollContainer from "../components/AppScrollContainer"
 import Button from "../components/Button"
-const Route = Routing.Route
-const Link = Routing.Link
+import Spacing from "../components/Spacing"
 
 export default class Home extends Component {
   render() {
     return (
       <AppScrollContainer backText="remove">
         <Button to="/login" title="Login" />
+        <Spacing size={15} />
         <Button to="/create" title="Create New User" />
       </AppScrollContainer>
     )